Refactor into transports and channels
